More naming fixes
