+++ cpp/dolfinx/mesh/CMakeLists.txt
@@ -2,11 +2,10 @@ set(HEADERS_mesh
   ${CMAKE_CURRENT_SOURCE_DIR}/dolfin_mesh.h
   ${CMAKE_CURRENT_SOURCE_DIR}/Mesh.h
   ${CMAKE_CURRENT_SOURCE_DIR}/Geometry.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/Partitioning.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/PermutationComputation.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/permutationcomputation.h
   ${CMAKE_CURRENT_SOURCE_DIR}/Topology.h
   ${CMAKE_CURRENT_SOURCE_DIR}/MeshTags.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/TopologyComputation.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/topologycomputation.h
   ${CMAKE_CURRENT_SOURCE_DIR}/cell_types.h
   ${CMAKE_CURRENT_SOURCE_DIR}/graphbuild.h
   ${CMAKE_CURRENT_SOURCE_DIR}/utils.h
@@ -15,7 +14,6 @@ set(HEADERS_mesh
 target_sources(dolfinx PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/Mesh.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/Geometry.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/Partitioning.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/PermutationComputation.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/Topology.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/TopologyComputation.cpp
